chore: custom resolvers for product references
